Confidence Interval stuff *ALMOST* complete.
TODO: Validate the CI


git-svn-id: http://svn.bramp.net:8080/svn/PhD/Work/userspace/threadnetperf@2054 e02701c3-087c-644a-a2a3-551e05f81084

+++ print.h
@@ -2,7 +2,7 @@
 
 void print_headers(const struct settings* settings);
 void print_results( const struct settings *settings, const struct stats *stats );
-void print_stats(unsigned long long sum,unsigned long long sumsquare,unsigned long long mean,double variance);
+void print_stats(float sum, float sumsquare, float mean, float variance) ;
 
 void print_hex(void *data, int size);
